Some Basic Player and Enemy Functionalities
Alien and Asteroid enemies have both had prefabs created. Enemies have stats associated with them. Only stat so far is health Enemies are destroyed if the health is at or below zero. Player movement has been set at a constant speed. There are three "lanes" in which the player can move. The Player can also shoot an object if there are 'enemy stats' on that object.