Player Stats and Alien collision
Player has three lives which decrease when the player comes in contact with an alien.
Aliens have 'Alien' tag, Asteroids have 'Asteroid' tag, Player has 'Player' tag. Player now also has a rigid body but doesn't use gravity. This is to get the collision detection to work. And the sphere collider on the Aliens are now a trigger.