Space Background
Imported some space background images from unity asset store. Background is parent of camera so it scrolls with the player. Materials are unlit shaders; stars are unlit/transparent and nebula is just unlit/texture. Applied the material to different quad object which are angled to be facing the camera. Texture is also continuously offset to give the impression of flying through space.