Alien and Player spaceships models imported
Alien model is now a low poly UFO. Player is now a low poly spaceship. Both still have their original collision object shapes. Player is still a cube collider and alien is still a sphere collider. Along with the obj files for the models, there are textures and materials with the models.