Health Display & Imported Text Mesh Pro
Needed text for UI so imported TextMeshPro. Also The Image for number of lives is an image of the player spaceship without a background. Also a count is next to this image, which is the amount of lives the player has. This is located at the top left of the screen. No functionality as of yet.