Damage Numbers
Played around with trying to use animations to show damage numbers but could only show one number per enemy damaged. Now when damaging enemies, new damageNumberPrefab is spawned and this hold a 3D text which moves upwards then disappears after 1 second. Functionality can be changed in the prefab with the script. For now though only the offset and speed  for the number to float upwards can be changed. Asteroid and Spaceship prefabs no longer use the canvas with text. Animations are useless now.