First Level Created and Movement changed
Made 2 more scenes, one is a scene template and the other is level 2. The scene that has been worked on the past commits is now level 1. Made some of the objects in level1 scene prefabs, so that I can use them in later scenes. ALSO movement had been changed to adding force to the player rigid body instead of using lane based movement.