Moved files, MainMenu scene, changed background
All the imported assets have been moved to a folder called Imports. Also, The main menu has been created. There is a play button and an  exit button. Exit closes the application and play switches the scene to the level 1 scene. Also, found a cool skybox so I've changed the level 1 background to the new skybox. Slowed down the stars on the scrolling background and got rid of the blue nebula scrolling background.