Game over overlay, made all the stats the same in all scenes
Game over overlay is now set to active at the start just so I can get a reference to the game object, then after is then set back to un-active, until the player dies. Then the overlay is set back to active. Changed the stats of the player in different scenes so its the same across the board. Also changed the level 2 scene skybox to the cool galaxy skybox